apps/worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "FastAPI worker for async tasks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "google-cloud-tasks>=2.16.0",
    "google-cloud-pubsub>=2.25.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "poethepoet>=0.32.0",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-custom-exit-code>=0.3.0",
]

[tool.uv]
managed = true

[tool.poe.tasks]
dev = "fastapi dev src/main.py --port 8080"
start = "fastapi run src/main.py --port 8080"
lint = "ruff check ."
format = "ruff format ."
check = ["lint", "format"]
typecheck = "mypy src"
test = "pytest tests/ --suppress-no-test-exit-code"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
